Moved MoneyDenominationTest and PlayerMoneyStateTest out of model package
